MBL-13150: Fixed issues in displaysBookmarkMenuItem tests
The displaysBookmarkMenuItem test in AssignmentDetailsRenderTest was failing regularly on APIs 27 and 28.  It looks like the root cause was that, for those APIs, the content description for the overflow menu can be some control characters in addition to "More options", and so we had to test for "contains 'More options'" rather than "is 'More options'" in a few places to get our logic working.

Along the way, I added some logic to enable a11y checking for ALL tests in AssignmentDetailsRenderTest, and fixed up a couple of a11y issues in the process.