Add AppLocalization strings to tests and used ally test instead
For web views, this was difficult as they register a different platform channel name depending on the id of the view created. The solution being pushed should provide enough of a general cover for letting web views display and pass in an accessibility test. Not sure why, but the androidTapTargetGuideline check is what causes these platform channels to need to be created.