[Flutter-Parent][MBL-13573] Add course details holder (#388)
* [Flutter-Parent][MBL-13573] Add course details holder

This is just the holder for the three tabs on the course details screen. Since those 3 screens all share the course and assignment data, this parent widget is loading that information for all of its children. For children to refresh the data, they need a consumer of the model, and just call model on the load. I’ve created 3 stub classes for these finer detail screens and shown how to do a refresh of data.

Also adds a new ‘BaseModel’ class, this is used for complex screens where we want to make use of provider for sharing data with children widgets. It keeps track of the loading state (via the ViewState property) and allows children and parents to use/load/refresh data when needed.