MBL-13558: Fix intermittent problems with discussion replies (#396)
Evidently, it can take a little time from when you reply to a discussion (or a reply) and when that reply shows up in the webview (and therefore our asserts for its existence pass).  Espresso does not naturally wait for this action, so I had to add a sleep().  And I found that 1 second was not long enough, so I made it 2 seconds.  The methods where these sleeps were added are called a total of 8 times by our interaction test suite, and once by our E2E test suite.

The sleep() is unfortunate, but I think it's necessary, and it's annoying when your PR is blocked on these tests failing.

If desired, I can create a ticket to find a way to create an idling resourse to wait for the webview to be updated.