[Student][Teacher][MBL-13725] Don’t filter notifications (#467)
Filtering notifications is problematic, as the domain users log in with may be different that the account’s root domain. This makes it hard to store notifications based on user info. Instead, we will just always show notifications.

Instead of keeping old communication channels alive, we now delete them when logging out or switching users. This prevents the currently logged in account from seeing notifications meant for an account which they don’t have access to.