[Flutter-Parent][MBL-13813] Course details grade fix
Clean up for LA-274 to include submissions as part of the assignment group call. This is getting priority fixed now as the api changed to return a list of submissions now, rather than a single submission for an assignment. Mostly, this is code cleanup to remove the extra submissions call.