[Flutter-Parent][MBL-13683] Report a problem dialog (#492)
* [Flutter-Parent][MBL-13683] Report a problem dialog

This lets app users contact support via the error report endpoint. Added to the ‘contact support’ button on crash screens. Also defaulted our ‘help’ button in the drawer to use this dialog to make it easier for testing. Should not be the default for the ‘help’ dialog in the end game.