[Student][Pact][MBL-13808] | pact round2 (#556)
* MBL-13808: Pact server-side work for Users/Enrollments/Courses

I had to upgrade from au.com.dius:pact-jvm-consumer-junit_2.12:3.5.12 to au.com.dius:pact-jvm-consumer-junit:4.0.2 to fix a problem with .numberType() not being specified correctly.  But this has also broken the way that we format query params in a request url, so stuff is still broken and I am stuck.  I have a slack message into Matt at Dius, and hopefully we can find a way to overcome this.  (Should probably upgrade from 4.0.2 to 4.0.5/latest when we're done.)

The upgrade caused a bunch of legacy Pact code to break (pact/canvas/parent, pact/canvas/teacher, pact/canvas/student, pact/docviewer).  Mechanical fixes were needed.

As to the latest stuff, it's mostly working minus the url query mis-formatting.  The one thing that needs to be worked out (or deleted) are the pact tests involving multiple courses -- that will take some surgery on the provider state side.

* MBL-13808: More tweaks to accomodate actual provider state