[Flutter Parent][E2E][MBL-13953] | data seeding foundation + dashboard E2E test (#596)
* MBL-13953: Initial commit to preserve changes

We can seed a user, but we'll probably want to do some refactoring to conform to canonical api usage.  Next: token-login and seeding a course for the user.

* MBL-13953: Another save-commit.

Can now seed users, courses and enrollments.  Changed app.test to dashboard,.test.  Added convenience logic for waiting for seeding to complete, grabbing seed objects.

* MBL-13953: Fleshed out an initial dashboard test

* MBL-13953: Clean-up

* MBL-13953: Addressed code formatting issues.

* MBL-13953: Code reformatting round 2

* MBL-13953: Addressed PR feedback: Removed a lot of commented-out lines.