Revert scroll change, prevent overflow error
Scrolling the entire screen has issues. Either we fail to keep the “Can’t find school” button on screen at all times, or we have nested scrolling. For now, just eliminating the warning on overflowing due to keyboard. Will have to revisit in the future for a more robust solution around small screen devices in landscape with a keyboard open.