MBL-14172: Fix RemoteConfigSettings E2E test issue
The problem was that "qr_login_enabled" is a prefix for "qr_login_enabled_teacher", so when I looked for the former (as a partial string) Espresso would get confused.  I fixed the matcher to look for rc_name+':'.

I also consolidated some common logic into a single method to get rid of some code duplication.