[Flutter-Parent] Fix course grade having no padding (#701)
To test:
The course Android Dev on our mobiledev sandbox fits the criteria for only showing the grade, without showing the grading period header. Before, the grade was right against the top of the screen. Now there is padding whether there is just the grade, just the grading period header, or both.