[Flutter-Parent] Fix opening iframe LTIs externally
Wasn’t opening an authenticated link before, so the user would have to login again.

To test:
Open some rich content with an lti in an iframe and click on the ‘Launch External Tool’ button at the bottom of the page. Verify it opens a web view directly to the content (no login required)