[Flutter-Parent] Fix issues with JNI linking for web views
For users that have a killed app and click on a link to a screen with a web view, then press the back button, a crash can occur. This sidesteps that _slightly_ by not displaying the web view in an inactive state, though hopefully one day the flutter team will fix the issue plaguing JNI linking and we won’t need this anymore.