[Flutter-Parent] Fix small, miscellaneous problems
When an empty url is passed to an image loader (FadeInImage/NetworkImage/etc) it can throw a non-fatal exception. It won’t crash the app but does create a non-fatal on Firebase.

Encoding domain provided by users so navigation can happen. If the url is invalid, the web login page may show an empty screen, but that’s what our native apps do in the same scenario (and it is better than not navigating to the screen with a non url safe character)

Updated native splash screen to use same accessible color as flutter