MBL-14436: Speculative fixes for some discussion tests
I ran the resulting DiscussionsInteractionTest suite against FTL several times and these changes seem to make a difference in stability.  There were some failures, followed by additional fixes, and it seems like we landed in a good place.

But I can't promise that this will be the last round of stabilization needed.