MBL-14436: Another round of stabilization for discussion tests (#839)
The basic problem was that we could press the "reply" link to reply to a reply, and the RCE editor would fail to pop up.  Or we could press the attachment icon in a reply and not be taken to the attachment.  So I did all that I could to improve those situations, and I *think* that they have improved.