MBL-14526: Fix Parent E2E tests
I was able to get rid of the "dart:ui" import in user_color.dart by changing the "color" field from Color to int, and making corresponding changes in other parts of the code.  (I *think* that the color field is actually transmitted as an int.)

It builds, the E2E tests pass, it looks right -- but I could use a review from someone more knowledgeable about how UserColor is transmitted / used.