[MBL-14642][All] Hide help buttons on landing pages
This is temporary while we figure out a better solution to mobile login
issues.

Support is overwhelmed with these requests right now and are
having to mass-reply to them all so this button is doing more harm than
good right now.

Test plan: Landing pages in all the apps should not have a help button.