MBL-14657: Converted StudentContextPageTest from Ditto
Unfortunately, converting to MockCanvas turned really ugly, as it wasn't readily apparent how to create the required response object.  So this test was converted to a "live" test instead.  It's not ideal, but it gets rid of our final use of Ditto in our test suite.

If we create an E2E suite for the Teacher app, we can make this test E2E.