[Teacher][MBL-14662] Fix viewing quiz files
SpeedGraderQuizWebViewFragment currently disallows internal routing. In the case of canvas file download links, this causes files to be loaded via the redirect URL which drops important information such as the file name and mime type. The result is that the file may download successfully, but the app does not know how to handle it.

This PR adds a route for context-less file download links (as seen in quiz file submissions), and updates the fragment to allow internal routing for files.

Additional changes:
 - Fixes a resource leak and typo in the media loader
 - Adds authentication to the quiz webview so users don't have to sign in again after switching logins
 - Centers the filename text in ViewUnsupportedFileFragment

refs: MBL-14662
affects: Teacher
release note: Fixed an issue where quiz files could not be viewed