MBL-14828: Disable animations during tests (#1074)
This is the magic bullet to disabling animations in the student app when running tests locally.  I don't know exactly what is happening in Flank/FTL (i.e., whether animations are being properly disabled there), but we don't appear to be having problems in FTL so I'm not going to worry about that.