[MBL-15391][Student][Teacher] Logout if the access token is invalid (#1290)
* [MBL-15391][Student][Teacher] Logout if the access token is invalid

refs: MBL-15391
affects: Student, Teacher
release note: Fixed a bug when the access is deleted the user would not be logged out of the app.

test plan: Login to either app, then delete the access token on the web (Account > Settings > Approved integrations). On startup the user should be directed to the login screen instead of the Dashboard. Test with QR login as well. Also smoke test masquerading.

* removed unused function

* fix UI test issues

* CR

* fix submodule tests