[MBL-14934][Parent] Observer alerts for missing courses cannot be opened
refs: MBL-14934
affects: Parent
release note: Fixed a bug where alerts may not load in certain cases.

test plan: Alerts with the locked_for_user attribute can not be opened. The icon of these assignments are changed to a lock and the title is indicates that they're locked. When trying to open when a snackbar should be visible.