[MBL-15695][Student] Dashboard announcements for K5 (#1430)
refs: MBL-15695
affects: Student
release note: Notifications are now visible on the elementary dashboard.

test plan: Check both the K5 and non-K5 Dashboard, both have been changed.

Create a course and invite the student, the notification card should be visible. Check accepting and declining the invitation.
Create a conference for a course, the notification card should be visible after starting the conference. Dismissing the conference once should keep the notification from popping up until the app data is cleared. Clicking the card should navigate to the conference in a custom tab.
Create a global announcement. All the announcement types should be visible with the correct color and icon. Clicking the announcement should take to the announcement details. Clearing the announcement should keep it from popping up EVER.