[MBL-15816][Student] Important dates (#1459)
refs: MBL-15816
affects: Student
release note: Added Important Dates to the Elementary Dashboard.

test plan:

On mobile the important dates tab should be visible.
On tablet in portrait the important dates should be in the side bar. The Schedule tab should load the mobile design.
On tablet in landscape the important dates should be in the side bar. The Schedule tab should load the old tablet design.
Both assignments and events should be visible.
Clicking an item should take to the details screen.
See specs for more info.