[MBL-16130][Student] Banner image on K5 Course details
refs: MBL-16130
affects: Student
release note: Added banner image to the Elementary Course Details screen.

test plan: 
Only the tablet layout is affected.
If the banner image has been set in the course settings it should show on the elementary course details screen.
If the banner image is not set, the course image is used instead.
If the course image and the banner image are not set the course color is used.