[MBL-17775][All] Force logout handling (#2922)
Test plan: Once we detect that our access token has expired we try to refresh it with our refresh token. If that fails we route the user to the login screen. After login the app should send the last api call with the new token and continue operations.

You can test it in one of two ways:

Setup the authentication plugin (see epic on how). The minimum duration is 2 days so it's not ideal for testing.
Go into account > Settings > Find the access token and delete it.
refs: MBL-17775
affects: Student, Teacher, Parent
release note: Improve expired session handling.